sway: use wireplumber components now
Waybar fixed the multi-display issues with wireplumber now so use that
module instead of pipewire.

Unset the idle_inhibitor when locking the screen so it actually turns
off the display after a while.

Move the waybar config into the proper bar section in sway config. Let
sway handle the background since we're using swaybg anyway.

Signed-off-by: Randolph Sapp <res.sapp@gmail.com>