all: restructure to use chezmoi
Restructure the current repository to use chezmoi. It's not perfect but
it's certainly more active than stow and addresses some of my gripes,
even if it introduces a few more on top.

Also remove the package tracking files for now. A chezmoi script can
replace them later.

Remove the bspwm config now that I no longer have a machine using it and
chezmoi prefers a single monolithic config structure.

Signed-off-by: Randolph Sapp <res.sapp@gmail.com>